Moving CustomTabBar to separate file
